requirements.txt
fastapi
uvicorn
python-telegram-bot
psycopg2-binary
pydantic
requests
python-dotenv
pytest
scikit-learn
prophet
